Development

+++ src/controllers/reviews.controller.js
@@ -101,10 +101,10 @@ export const getProductReview = async (req, res) => {
   const { productId } = req.params;
 
   try {
-    const reviews = await Review.find({ product: productId }).populate(
-      "user",
-      "-password"
-    );
+    const reviews = await Review.find({
+      product: productId,
+      published: true,
+    }).populate("user", "-password");
     if (reviews) {
       const response = {
         success: true,
@@ -166,14 +166,64 @@ export const updateProductReview = async (req, res) => {
     return res.status(500).json(response);
   }
 };
+
+export const updateProductReviewStatus = async (req, res) => {
+  const { user, reviewId } = req.params;
+  try {
+    const existreview = await Review.findById(reviewId);
+    if (!existreview) {
+      const response = {
+        success: true,
+        statuscode: 400,
+        message: "Review not found.",
+      };
+      res.status(400).json(response);
+    }
+    const review = await Review.findByIdAndUpdate(
+      reviewId,
+      { published: !existreview.published },
+      {
+        new: true,
+        runValidators: true,
+      }
+    );
+    if (!review) {
+      const response = {
+        success: true,
+        statuscode: 400,
+        message: "Review not found.",
+      };
+      res.status(400).json(response);
+    }
+    const response = {
+      success: true,
+      statuscode: 201,
+      data: review,
+      message: "status updated.",
+    };
+    return res.status(201).json(response);
+  } catch (error) {
+    const response = {
+      success: true,
+      statuscode: 500,
+      message: "An error occurred while updating the review.",
+    };
+    res.status(500).json(response);
+  }
+};
 export const deleteProductReview = async (req, res) => {
   const { reviewId } = req.params;
 
   try {
-    const review = await Review.findByIdAndDelete(reviewId);
-    if (!review) {
-      return res.status(404).json({ message: "Review not found." });
+    const reviews = await Review.findById(reviewId);
+    if (!reviews) {
+      return res.status(404).json({
+        success: true,
+        statuscode: 404,
+        message: "Review not found.",
+      });
     }
+    const review = await Review.findByIdAndDelete(reviewId);
 
     // Remove the review ID from the associated product's reviews array
     const product = await Product.findById(review.product);
@@ -183,7 +233,12 @@ export const deleteProductReview = async (req, res) => {
       await product.save();
     }
 
-    res.json({ message: "Review deleted successfully." });
+    const response = {
+      success: true,
+      statuscode: 200,
+      message: "review removed.",
+    };
+    res.status(200).json(response);
   } catch (error) {
     const response = {
       success: true,
